rebalance策略避免重复

+++ d10-scheduler-registry-center/src/main/java/com/github/krystalics/d10/scheduler/registry/service/impl/RebalanceServiceImpl.java
@@ -1,9 +1,9 @@
 package com.github.krystalics.d10.scheduler.registry.service.impl;
 
 import com.github.krystalics.d10.scheduler.common.constant.CommonConstants;
-import com.github.krystalics.d10.scheduler.registry.service.RebalanceService;
 import com.github.krystalics.d10.scheduler.dao.mapper.TaskMapper;
 import com.github.krystalics.d10.scheduler.dao.qm.TaskQM;
+import com.github.krystalics.d10.scheduler.registry.service.RebalanceService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.zookeeper.CreateMode;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,31 +26,42 @@ public class RebalanceServiceImpl implements RebalanceService {
     @Autowired
     private ZookeeperServiceImpl zookeeperService;
 
+    /**
+     * 当新leader替换老leader上位，就会在startRunner和live node delete重复执行 rebalance
+     * 所以需要去重
+     */
+    private String rebalanceAddress = "";
+
     @Override
-    public void rebalance(String address) throws Exception {
-        log.info("scheduler system begin rebalanced!");
-        log.info("1.to create /shard node");
-        zookeeperService.createNodeIfNotExist(CommonConstants.ZK_SHARD_NODE, address, CreateMode.EPHEMERAL);
-        log.info("2.assign the task to schedulers");
+    public synchronized void rebalance(String address) throws Exception {
+        if (!rebalanceAddress.equals(address)) {
+            rebalanceAddress = address;
+            log.info("scheduler system begin rebalanced!");
+            log.info("1.to create /shard node");
+            zookeeperService.createNodeIfNotExist(CommonConstants.ZK_SHARD_NODE, address, CreateMode.EPHEMERAL);
+            log.info("2.assign the task to schedulers");
 
-        final List<String> liveNodes = zookeeperService.liveNodes();
-        log.info("rebalancing ,all live nodes are {}", liveNodes);
-        TaskQM taskQM = new TaskQM();
-        taskQM.setState(2);
-        final int taskSize = taskMapper.count(taskQM);
-        final int nodeSize = liveNodes.size();
+            final List<String> liveNodes = zookeeperService.liveNodes();
+            log.info("rebalancing ,all live nodes are {}", liveNodes);
+            TaskQM taskQM = new TaskQM();
+            taskQM.setState(2);
+            final int taskSize = taskMapper.count(taskQM);
+            final int nodeSize = liveNodes.size();
 
-        int step = 0;
-        final int stepLong = taskSize / nodeSize;
+            int step = 0;
+            final int stepLong = taskSize / nodeSize;
 
-        for (String liveNode : liveNodes) {
-            step += stepLong;
+            for (String liveNode : liveNodes) {
+                step += stepLong;
 //            ClusterInfo.setNodeRange(liveNode, step);
+            }
+
+            log.info("3.delete the /shard node");
+            zookeeperService.deleteNode(CommonConstants.ZK_SHARD_NODE);
+            rebalanceAddress = "";
+        }else {
+            log.warn("the former one has been executed, node = {}",address);
         }
-
-        log.info("3.delete the /shard node");
-        zookeeperService.deleteNode(CommonConstants.ZK_SHARD_NODE);
-
     }
 
     public void shard(String address) throws Exception {
